# doc/api.yaml
openapi: 3.0.0
info:
  title: Home Library Service
  description: Home music library service
  version: 1.0.0

servers:
  - url: / # Our NestJS app serves from the root

components:
  schemas:
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: User's unique identifier
        login:
          type: string
          example: 'TestUser'
        version:
          type: number
          example: 1
          description: Version of the user record, increments on update
        createdAt: # Corrected from createAt
          type: number
          example: 1655000000
          description: Timestamp of creation
        updatedAt: # Corrected from updateAt
          type: number
          example: 1655000000
          description: Timestamp of last update
      required:
        - id
        - login
        - version
        - createdAt
        - updatedAt

    CreateUserDto:
      type: object
      properties:
        login:
          type: string
        password:
          type: string
      required:
        - login
        - password

    UpdatePasswordDto:
      type: object
      properties:
        oldPassword:
          type: string
        newPassword:
          type: string
      required:
        - oldPassword
        - newPassword

    LoginDto:
      type: object
      properties:
        login:
          type: string
        password:
          type: string
      required:
        - login
        - password

    LoginResponseDto:
      type: object
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string
      required:
        - accessToken
        - refreshToken

    Artist:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          example: 'Freddie Mercury'
        grammy:
          type: boolean
          example: false
      required:
        - id
        - name
        - grammy

    CreateArtistDto:
      type: object
      properties:
        name:
          type: string
        grammy:
          type: boolean
      required:
        - name
        - grammy

    UpdateArtistDto:
      type: object
      properties:
        name:
          type: string
        grammy:
          type: boolean
      required:
        - name
        - grammy

    Album:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          example: 'Innuendo'
        year:
          type: integer
          example: 1991
        artistId:
          type: string
          format: uuid
          nullable: true
      required:
        - id
        - name
        - year

    CreateAlbumDto:
      type: object
      properties:
        name:
          type: string
        year:
          type: integer
        artistId:
          type: string
          format: uuid
          nullable: true
      required:
        - name
        - year

    UpdateAlbumDto:
      type: object
      properties:
        name:
          type: string
        year:
          type: integer
        artistId:
          type: string
          format: uuid
          nullable: true
      required:
        - name
        - year

    Track:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          example: The Show Must Go On
        artistId:
          type: string
          format: uuid
          nullable: true
        albumId:
          type: string
          format: uuid
          nullable: true
        duration:
          type: integer
          description: In seconds
          example: 262
      required:
        - id
        - name
        - duration

    CreateTrackDto:
      type: object
      properties:
        name:
          type: string
        artistId:
          type: string
          format: uuid
          nullable: true
        albumId:
          type: string
          format: uuid
          nullable: true
        duration:
          type: integer
      required:
        - name
        - duration

    UpdateTrackDto:
      type: object
      properties:
        name:
          type: string
        artistId:
          type: string
          format: uuid
          nullable: true
        albumId:
          type: string
          format: uuid
          nullable: true
        duration:
          type: integer
      required:
        - name
        - duration

    FavoritesResponse: # Corrected name from Favorites to match assignment
      type: object
      properties:
        artists:
          type: array
          items:
            $ref: '#/components/schemas/Artist'
        albums:
          type: array
          items:
            $ref: '#/components/schemas/Album'
        tracks:
          type: array
          items:
            $ref: '#/components/schemas/Track'
      required:
        - artists
        - albums
        - tracks

    MessageResponse: # For simple success messages
      type: object
      properties:
        message:
          type: string
          example: 'Operation successful'

  responses:
    UnauthorizedError:
      description: Access token is missing or invalid
      content:
        application/json:
          schema:
            type: object
            properties:
              statusCode:
                type: integer
                example: 401
              message:
                type: string
                example: Unauthorized
    ForbiddenError:
      description: Access forbidden
      content:
        application/json:
          schema:
            type: object
            properties:
              statusCode:
                type: integer
                example: 403
              message:
                type: string
                example: Forbidden
    NotFoundError:
      description: Resource not found
      content:
        application/json:
          schema:
            type: object
            properties:
              statusCode:
                type: integer
                example: 404
              message:
                type: string
                example: Not Found
    BadRequestError:
      description: Bad request (invalid data or invalid UUID)
      content:
        application/json:
          schema:
            type: object
            properties:
              statusCode:
                type: integer
                example: 400
              message:
                type: array # Or string, depending on ValidationPipe output
                items:
                  type: string
                example: ['password should not be empty'] # Example for ValidationPipe
              error:
                type: string
                example: Bad Request
    UnprocessableEntityError:
      description: Unprocessable Entity (e.g., referenced entity does not exist)
      content:
        application/json:
          schema:
            type: object
            properties:
              statusCode:
                type: integer
                example: 422
              message:
                type: string
                example: 'Artist with id ... not found' # Example
              error:
                type: string
                example: Unprocessable Entity
    ConflictError:
      description: Conflict (e.g., resource already exists)
      content:
        application/json:
          schema:
            type: object
            properties:
              statusCode:
                type: integer
                example: 409
              message:
                type: string
                example: 'User with this login already exists'
              error:
                type: string
                example: Conflict

  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

security: # Global security setting, applies to all paths unless overridden
  - bearerAuth: []

paths:
  /auth/signup:
    post:
      tags:
        - Auth
      security: [] # Public endpoint
      summary: Sign up a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserDto'
      responses:
        '201':
          description: User successfully created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User' # Returns the created user (without password)
        '400':
          $ref: '#/components/responses/BadRequestError'
        '409':
          $ref: '#/components/responses/ConflictError' # User with this login already exists

  /auth/login:
    post:
      tags:
        - Auth
      security: [] # Public endpoint
      summary: Log in an existing user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginDto'
      responses:
        '200':
          description: Successful login, returns access and refresh tokens
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponseDto'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401': # Changed from 403 as per typical login failure
          description: Login failed (user not found or password incorrect)
          content:
            application/json:
              schema:
                type: object
                properties:
                  statusCode:
                    type: integer
                    example: 401
                  message:
                    type: string
                    example: Unauthorized

  /auth/refresh:
    post:
      tags:
        - Auth
      security: [] # This endpoint handles its own token logic, often public or using refresh token as auth
      summary: Refresh access and refresh tokens
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                refreshToken:
                  type: string
              required:
                - refreshToken
      responses:
        '200':
          description: Tokens successfully refreshed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponseDto'
        '401':
          description: Refresh token missing or malformed
          content:
            application/json:
              schema:
                type: object
                properties:
                  statusCode:
                    type: integer
                    example: 401
                  message:
                    type: string
                    example: Unauthorized
        '403':
          description: Refresh token invalid or expired
          content:
            application/json:
              schema:
                type: object
                properties:
                  statusCode:
                    type: integer
                    example: 403
                  message:
                    type: string
                    example: Forbidden

  /user: # Path changed from /users
    get:
      tags:
        - User
      summary: Get all users
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/UnauthorizedError'

  /user/{id}: # Path parameter changed from userId to id
    parameters:
      - name: id
        in: path
        required: true
        description: User ID
        schema:
          type: string
          format: uuid
    get:
      tags:
        - User
      summary: Get single user by id
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          $ref: '#/components/responses/BadRequestError' # Description: Bad request. id is invalid (not uuid)
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError' # Description: User not found
    put:
      tags:
        - User
      summary: Update user's password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdatePasswordDto'
      responses:
        '200':
          description: Password successfully updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User' # Returns updated user
        '400':
          $ref: '#/components/responses/BadRequestError' # Description: Bad request. id is invalid (not uuid) or DTO invalid
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError' # Description: oldPassword is wrong
        '404':
          $ref: '#/components/responses/NotFoundError' # Description: User not found
    delete:
      tags:
        - User
      summary: Delete user
      responses:
        '204':
          description: The user has been deleted
        '400':
          $ref: '#/components/responses/BadRequestError' # Description: Bad request. id is invalid (not uuid)
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError' # Description: User not found

  /artist: # Path changed
    get:
      tags:
        - Artist
      summary: Get all artists
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Artist'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
    post:
      tags:
        - Artist
      summary: Add new artist
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateArtistDto'
      responses:
        '201':
          description: Artist created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Artist'
        '400':
          $ref: '#/components/responses/BadRequestError' # Description: Bad request. body does not contain required fields
        '401':
          $ref: '#/components/responses/UnauthorizedError'

  /artist/{id}: # Path parameter changed
    parameters:
      - name: id
        in: path
        required: true
        description: Artist ID
        schema:
          type: string
          format: uuid
    get:
      tags:
        - Artist
      summary: Get single artist by id
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Artist'
        '400':
          $ref: '#/components/responses/BadRequestError' # Description: Bad request. id is invalid (not uuid)
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError' # Description: Artist not found
    put:
      tags:
        - Artist
      summary: Update artist information
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateArtistDto'
      responses:
        '200':
          description: Artist updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Artist'
        '400':
          $ref: '#/components/responses/BadRequestError' # Description: Bad request. id is invalid (not uuid) or DTO invalid
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError' # Description: Artist not found
    delete:
      tags:
        - Artist
      summary: Delete artist
      responses:
        '204':
          description: Artist deleted
        '400':
          $ref: '#/components/responses/BadRequestError' # Description: Bad request. id is invalid (not uuid)
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError' # Description: Artist not found

  /album: # Path changed
    get:
      tags:
        - Album
      summary: Get all albums
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Album'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
    post:
      tags:
        - Album
      summary: Add new album
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAlbumDto'
      responses:
        '201':
          description: Album created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Album'
        '400':
          $ref: '#/components/responses/BadRequestError' # Description: Bad request. body does not contain required fields (or referenced artistId not found)
        '401':
          $ref: '#/components/responses/UnauthorizedError'

  /album/{id}: # Path parameter changed
    parameters:
      - name: id
        in: path
        required: true
        description: Album ID
        schema:
          type: string
          format: uuid
    get:
      tags:
        - Album
      summary: Get single album by id
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Album'
        '400':
          $ref: '#/components/responses/BadRequestError' # Description: Bad request. id is invalid (not uuid)
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError' # Description: Album not found
    put:
      tags:
        - Album
      summary: Update album information
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateAlbumDto'
      responses:
        '200':
          description: Album updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Album'
        '400':
          $ref: '#/components/responses/BadRequestError' # Description: Bad request. id is invalid (not uuid) or DTO invalid (or referenced artistId not found)
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError' # Description: Album not found
    delete:
      tags:
        - Album
      summary: Delete album
      responses:
        '204':
          description: Album deleted
        '400':
          $ref: '#/components/responses/BadRequestError' # Description: Bad request. id is invalid (not uuid)
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError' # Description: Album not found

  /track: # Path changed
    get:
      tags:
        - Track
      summary: Get all tracks
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Track'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
    post:
      tags:
        - Track
      summary: Add new track
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTrackDto'
      responses:
        '201':
          description: Track created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Track'
        '400':
          $ref: '#/components/responses/BadRequestError' # Description: Bad request. body does not contain required fields (or referenced artistId/albumId not found)
        '401':
          $ref: '#/components/responses/UnauthorizedError'

  /track/{id}: # Path parameter changed
    parameters:
      - name: id
        in: path
        required: true
        description: Track ID
        schema:
          type: string
          format: uuid
    get:
      tags:
        - Track
      summary: Get single track by id
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Track'
        '400':
          $ref: '#/components/responses/BadRequestError' # Description: Bad request. id is invalid (not uuid)
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError' # Description: Track not found
    put:
      tags:
        - Track
      summary: Update track information
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTrackDto'
      responses:
        '200':
          description: Track updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Track'
        '400':
          $ref: '#/components/responses/BadRequestError' # Description: Bad request. id is invalid (not uuid) or DTO invalid (or referenced artistId/albumId not found)
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError' # Description: Track not found
    delete:
      tags:
        - Track
      summary: Delete track
      responses:
        '204':
          description: Track deleted
        '400':
          $ref: '#/components/responses/BadRequestError' # Description: Bad request. id is invalid (not uuid)
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError' # Description: Track not found

  /favs:
    get:
      tags:
        - Favorites
      summary: Get all favorites
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FavoritesResponse' # Corrected schema name
        '401':
          $ref: '#/components/responses/UnauthorizedError'

  /favs/track/{id}:
    parameters:
      - name: id
        in: path
        required: true
        description: Track ID to add/remove from favorites
        schema:
          type: string
          format: uuid
    post:
      tags:
        - Favorites
      summary: Add track to favorites
      responses:
        '201':
          description: Track added to favorites
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse' # Example: {"message": "Track added"}
        '400':
          $ref: '#/components/responses/BadRequestError' # Description: id is invalid (not uuid)
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '422':
          $ref: '#/components/responses/UnprocessableEntityError' # Description: Track with id doesn't exist in library
    delete:
      tags:
        - Favorites
      summary: Delete track from favorites
      responses:
        '204':
          description: Track deleted from favorites
        '400':
          $ref: '#/components/responses/BadRequestError' # Description: id is invalid (not uuid)
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError' # Description: Track not found in favorites

  /favs/album/{id}:
    parameters:
      - name: id
        in: path
        required: true
        description: Album ID to add/remove from favorites
        schema:
          type: string
          format: uuid
    post:
      tags:
        - Favorites
      summary: Add album to favorites
      responses:
        '201':
          description: Album added to favorites
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '422':
          $ref: '#/components/responses/UnprocessableEntityError' # Description: Album with id doesn't exist in library
    delete:
      tags:
        - Favorites
      summary: Delete album from favorites
      responses:
        '204':
          description: Album deleted from favorites
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError' # Description: Album not found in favorites

  /favs/artist/{id}:
    parameters:
      - name: id
        in: path
        required: true
        description: Artist ID to add/remove from favorites
        schema:
          type: string
          format: uuid
    post:
      tags:
        - Favorites
      summary: Add artist to favorites
      responses:
        '201':
          description: Artist added to favorites
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '422':
          $ref: '#/components/responses/UnprocessableEntityError' # Description: Artist with id doesn't exist in library
    delete:
      tags:
        - Favorites
      summary: Delete artist from favorites
      responses:
        '204':
          description: Artist deleted from favorites
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError' # Description: Artist not found in favorites